requirements.txt
absl-py==1.0.0
aiobotocore==2.4.0
aiohttp==3.8.1
aiohttp-retry==2.8.3
aioitertools==0.10.0
aiosignal==1.2.0
alembic==1.7.6
amqp==5.1.1
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
async-timeout==4.0.2
asyncssh==2.12.0
atpublic==3.1.1
attrs==22.1.0
autopage==0.5.0
awscli==1.25.69
backcall==0.2.0
billiard==3.6.4.0
bleach==4.1.0
boto3==1.24.59
botocore==1.27.68
cachetools==5.0.0
celery==5.2.7
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cliff==3.10.1
cmaes==0.8.2
cmd2==2.4.0
colorama==0.4.4
colorlog==6.6.0
commonmark==0.9.1
configobj==5.0.6
cramjam==2.5.0
cryptography==38.0.1
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
dictdiffer==0.9.0
diskcache==5.4.0
distro==1.7.0
docutils==0.16
dpath==2.0.6
dulwich==0.20.46
dvc==2.24.0
dvc-data==0.4.0
dvc-http==2.19.1
dvc-objects==0.2.0
dvc-render==0.0.9
dvc-s3==2.19.0
dvc-task==0.1.2
dvclive==0.10.0
entrypoints==0.4
executing==0.8.3
fastjsonschema==2.16.1
flatten-dict==0.4.2
flufl.lock==7.1.1
fonttools==4.29.1
frozenlist==1.3.0
fsspec==2022.8.2
funcy==1.17
future==0.18.2
gitdb==4.0.9
GitPython==3.1.27
google-auth==2.6.0
google-auth-oauthlib==0.4.6
grandalf==0.6
greenlet==1.1.2
grpcio==1.44.0
hydra-core==1.2.0
idna==3.3
importlib-metadata==4.12.0
ipykernel==6.9.1
ipython==8.1.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.1
Jinja2==3.0.3
jmespath==1.0.1
joblib==1.1.0
jsonschema==4.15.0
jupyter==1.0.0
jupyter-client==7.1.2
jupyter-console==6.4.3
jupyter-core==4.11.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kiwisolver==1.3.2
kombu==5.2.4
lightgbm==3.3.2
llvmlite==0.38.1
Mako==1.1.6
Markdown==3.3.6
MarkupSafe==2.1.0
matplotlib==3.5.2
matplotlib-inline==0.1.3
mistune==0.8.4
multidict==6.0.2
nanotime==0.5.2
nbclient==0.5.12
nbconvert==6.4.2
nbformat==5.4.0
nest-asyncio==1.5.4
networkx==2.8.6
notebook==6.4.8
numba==0.55.2
numpy==1.22.4
oauthlib==3.2.0
omegaconf==2.2.3
optuna==3.0.0
packaging==21.3
pandas==1.4.1
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.9.0
patsy==0.5.2
pbr==5.8.1
pickleshare==0.7.5
Pillow==9.0.1
plotly==5.7.0
prettytable==3.2.0
prometheus-client==0.13.1
prompt-toolkit==3.0.28
protobuf==3.19.4
psutil==5.9.2
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyDeprecate==0.3.1
pydot==1.4.2
pygit2==1.10.1
Pygments==2.11.2
pygtrie==2.5.0
pyparsing==3.0.7
pyperclip==1.8.2
pyreadline3==3.4.1
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2021.3
pywin32==303
pywinpty==2.0.5
PyYAML==5.4.1
pyzmq==22.3.0
qtconsole==5.2.2
QtPy==2.0.1
requests==2.27.1
requests-oauthlib==1.3.1
rich==12.5.1
rsa==4.7.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
s3fs==2022.8.2
s3transfer==0.6.0
scikit-learn==1.0.2
scipy==1.8.0
scmrepo==0.0.25
seaborn==0.11.2
Send2Trash==1.8.0
setuptools-scm==7.0.5
shortuuid==1.0.9
shtab==1.5.5
six==1.16.0
smmap==5.0.0
SQLAlchemy==1.4.32
stack-data==0.2.0
statsmodels==0.13.2
stevedore==3.5.0
tabulate==0.8.10
tenacity==8.0.1
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
terminado==0.13.3
testpath==0.6.0
threadpoolctl==3.1.0
tomli==2.0.1
tomlkit==0.11.4
tornado==6.1
tqdm==4.64.1
traitlets==5.3.0
typing_extensions==4.3.0
urllib3==1.26.8
vine==5.0.0
voluptuous==0.13.1
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.3
widgetsnbextension==3.5.2
wincertstore==0.2
wrapt==1.14.1
yarl==1.7.2
zc.lockfile==2.0
zipp==3.8.1
